+++ src/Cwaff-Actives/BulletbotImplant.cs
@@ -4,8 +4,10 @@ public class BulletbotImplant : CwaffActive
 {
     public static string ItemName         = "Bulletbot Implant";
     public static string ShortDescription = "Loyal Gunpanions"; //How to Train your Dragun
-    public static string LongDescription  = "Grants the nearest unarmed companion a random gun and the ability to fire it at nearby enemies.";
+    public static string LongDescription  = "Grants the nearest unarmed companion a random gun and the ability to fire it at nearby enemies. Quadruples the chance of finding companions while held.";
     public static string Lore             = "A microchip designed to rewire the neurons and DNA of its host to instantaneously train them in the art of armed combat. Conventional wisdom posits that if you give a dog a gun, he'll bark at it all day, but if you teach a dog to gun, he'll fight by your side for the rest of his life.";
+
+    private const float _COMPANION_CHANCE_MULT = 4f;
 
     private AIActor _nearestCompanion = null;
 
@@ -20,6 +22,13 @@ public class BulletbotImplant : CwaffActive
         item.AddToShop(ModdedShopType.Handy);
 
         FakeItem.Create<UsedBulletbotImplant>();
+    }
+
+    public override void OnFirstPickup(PlayerController player)
+    {
+        base.OnFirstPickup(player);
+        foreach (var companion in Cammy._CompanionTable.defaultItemDrops.elements)
+            this.IncreaseLootChance(companion.pickupId, _COMPANION_CHANCE_MULT);
     }
 
     public override bool CanBeUsed(PlayerController user)